// react-module/src/openapi/models/pack-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PackBasicDto
 */
export interface PackBasicDto {
    /**
     * 
     * @type {number}
     * @memberof PackBasicDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof PackBasicDto
     */
    barcodeFromId?: any;
    /**
     * 
     * @type {number}
     * @memberof PackBasicDto
     */
    weightNet?: any;
    /**
     * 
     * @type {number}
     * @memberof PackBasicDto
     */
    weightGross?: any;
    /**
     * 
     * @type {string}
     * @memberof PackBasicDto
     */
    deliveryData?: any;
    /**
     * 
     * @type {string}
     * @memberof PackBasicDto
     */
    packStatus?: PackBasicDtoPackStatusEnum;
    /**
     * 
     * @type {ContractorDto}
     * @memberof PackBasicDto
     */
    contractor?: any;
    /**
     * 
     * @type {InventoryDto}
     * @memberof PackBasicDto
     */
    inventory?: any;
    /**
     * 
     * @type {string}
     * @memberof PackBasicDto
     */
    inventoryString?: any;
    /**
     * 
     * @type {boolean}
     * @memberof PackBasicDto
     */
    createdByMobileApp?: any;
    /**
     * 
     * @type {PackDimension}
     * @memberof PackBasicDto
     */
    dimension?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum PackBasicDtoPackStatusEnum {
    INIT = 'INIT',
    CONFIRMED = 'CONFIRMED',
    PACKED = 'PACKED',
    SENT = 'SENT'
}